+++ src/main/java/lermitage/intellij/extra/icons/enablers/AbstractInFolderEnabler.java
@@ -9,12 +9,15 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.search.FilenameIndex;
 import com.intellij.psi.search.GlobalSearchScope;
+import dev.failsafe.Failsafe;
+import dev.failsafe.RetryPolicy;
 import lermitage.intellij.extra.icons.Globals;
 import lermitage.intellij.extra.icons.cfg.SettingsService;
 import lermitage.intellij.extra.icons.utils.ProjectUtils;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
+import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
@@ -26,6 +29,9 @@ public abstract class AbstractInFolderEnabler implements IconEnabler {
     private static final Logger LOGGER = Logger.getInstance(AbstractInFolderEnabler.class);
     private final String className = this.getClass().getSimpleName();
 
+    private final int FILENAME_INDEX_QUERY_MAX_ATTEMPTS = 3;
+    private final int FILENAME_INDEX_QUERY_RETRY_DELAY_MS = 60;
+
     private boolean initialized = false;
     private boolean indexErrorReported = false;
     protected Set<String> folders = Collections.emptySet();
@@ -35,27 +41,43 @@ public abstract class AbstractInFolderEnabler implements IconEnabler {
     /** The name of this icon enabler. Used to identify disabled icon enabler if an error occurred. */
     public abstract String getName();
 
-    protected synchronized void init(@NotNull Project project) {
+    @Override
+    public synchronized void init(@NotNull Project project) {
         long t1 = System.currentTimeMillis();
         String[] filenamesToSearch = getFilenamesToSearch();
         Collection<VirtualFile> virtualFilesByName;
         try {
-            // TODO migrate to getVirtualFilesByName(getFilenamesToSearch()[0], true, GlobalSearchScope.projectScope(project))
-            //  in 2023 and set minimal IDE version to 2022.1 (221)
-            virtualFilesByName = FilenameIndex.getVirtualFilesByName(
-                project,
-                getFilenamesToSearch()[0],
-                true,
-                GlobalSearchScope.projectScope(project));
+            RetryPolicy<Collection<VirtualFile>> retryPolicy = RetryPolicy.<Collection<VirtualFile>>builder()
+                .handle(Exception.class)
+                .withMaxAttempts(FILENAME_INDEX_QUERY_MAX_ATTEMPTS)
+                .onRetry(event -> LOGGER.warn(getName() + " IconEnabler failed to query IDE filename index " +
+                    "(attempt " + event.getAttemptCount() + "/" + FILENAME_INDEX_QUERY_MAX_ATTEMPTS + "). " +
+                    "Will try again in " + FILENAME_INDEX_QUERY_RETRY_DELAY_MS + " ms. " + event))
+                .withDelay(Duration.ofMillis(FILENAME_INDEX_QUERY_RETRY_DELAY_MS)).build();
+
+            virtualFilesByName = Failsafe.with(retryPolicy).get(() -> {
+                // TODO migrate to getVirtualFilesByName(getFilenamesToSearch()[0], true, GlobalSearchScope.projectScope(project))
+                //  in 2023 and set minimal IDE version to 2022.1 (221)
+                return FilenameIndex.getVirtualFilesByName(
+                    project,
+                    getFilenamesToSearch()[0],
+                    true,
+                    GlobalSearchScope.projectScope(project));
+            });
         } catch (Exception e) {
             initialized = true;
             if (!indexErrorReported) {
                 indexErrorReported = true;
-                String msg = "Failed to query IDE filename index. <b>This feature won't work: " + getName() + "</b>. " +
-                    "If this is the first time you see this message, and if you really need this feature, " +
+                String msg = "Failed to query IDE filename index. <b>This feature won't work: " + getName() + "</b>.<br>" +
+                    "📢 If this is the first time you see this message, and if you really need this feature, " +
                     "please restart your IDE. If it doesn't help, try to clear the file system cache and " +
-                    "Local History (go to File, Invalidate Caches...).<br><hr>" +
-                    "To disable this notification, please go to <b>File</b>, <b>Settings...</b>, <b>Extra Icons</b>, and check <b>Ignore plugin's warnings</b>.";
+                    "Local History (go to <i>File</i>, <i>Invalidate Caches...</i>).<br>" +
+                    "📢 You can also wait until indexing is done, then go to <i>File</i>, <i>Settings</i>, <i>Appearance &amp; " +
+                    "Behavior</i>, <i>Extra Icons</i>, and hit the <b>Reload projects icons</b> button.<br><hr>" +
+                    "<b>It seems to be a JetBrains issue</b> (feel free to <b>upvote</b> " +
+                    "https://youtrack.jetbrains.com/issue/IDEA-289822). Please do not open a new ticket for that.<br>" +
+                    "To disable this notification, please go to <i>File</i>, <i>Settings</i>, <i>Extra Icons</i>, and " +
+                    "check <b>Ignore plugin's warnings</b>.";
                 LOGGER.warn(msg, e);
                 if (!SettingsService.getIDEInstance().getIgnoreWarnings()) {
                     NotificationGroupManager.getInstance().getNotificationGroup(Globals.PLUGIN_GROUP_DISPLAY_ID)